# src/ClientApiSdkGen/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SMEIoT API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:5001"
    },
    {
      "url": "https://smeiot.se"
    }
  ],
  "paths": {
    "/api/admin/users": {
      "get": {
        "tags": [
          "AdminUsers"
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "roles",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserApiModelList"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{userName}": {
      "get": {
        "tags": [
          "AdminUsers"
        ],
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserApiModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sensors/{name}/users": {
      "post": {
        "tags": [
          "SensorAssignments"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserSensorBindingModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorAssignmentApiModel"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "SensorAssignments"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSensorIAsyncEnumerable"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sensors": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorDetailsApiModelList"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sensors"
        ],
        "requestBody": {
          "content": {
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/SensorLocatorBindingModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicSensorApiModel"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/sensors/{deviceName}/{sensorName}": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "parameters": [
          {
            "name": "deviceName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sensorName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorDetailsApiModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "requestBody": {
          "content": {
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/ValidatedUserCredentialsBindingModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicUserApiModel"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userName}/password": {
      "put": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmedUserCredentialsUpdateBindingModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserCredentialsUpdateApiModel"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userName}": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicUserApiModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdminUserApiModel": {
        "required": [
          "createdAt",
          "id",
          "lastSeenAt",
          "roles",
          "userName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "lastSeenAt": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "AdminUserApiModelList": {
        "required": [
          "total",
          "users"
        ],
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserApiModel"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": false
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "UserCredentialsUpdateApiModel": {
        "required": [
          "createdAt",
          "passwordUpdated",
          "roles",
          "userName"
        ],
        "type": "object",
        "properties": {
          "passwordUpdated": {
            "type": "boolean"
          },
          "userName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "BasicSensorApiModel": {
        "required": [
          "deviceName",
          "sensorName"
        ],
        "type": "object",
        "properties": {
          "sensorName": {
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "AssignUserSensorBindingModel": {
        "required": [
          "userName"
        ],
        "type": "object",
        "properties": {
          "userName": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SensorAssignmentApiModel": {
        "required": [
          "assignmentCount",
          "name",
          "userName"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "assignmentCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "UserSensorIAsyncEnumerable": {
        "type": "object",
        "additionalProperties": false
      },
      "NumberTimeSeriesApiModel": {
        "required": [
          "createdAt",
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SensorDetailsApiModel": {
        "required": [
          "data",
          "deviceName",
          "sensorName",
          "startedAt"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NumberTimeSeriesApiModel"
            }
          },
          "startedAt": {
            "type": "string"
          },
          "sensorName": {
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SensorDetailsApiModelList": {
        "required": [
          "sensors"
        ],
        "type": "object",
        "properties": {
          "sensors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SensorDetailsApiModel"
            }
          }
        },
        "additionalProperties": false
      },
      "SensorLocatorBindingModel": {
        "required": [
          "deviceName",
          "name"
        ],
        "type": "object",
        "properties": {
          "deviceName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ValidatedUserCredentialsBindingModel": {
        "required": [
          "password",
          "userName"
        ],
        "type": "object",
        "properties": {
          "userName": {
            "maxLength": 1000,
            "minLength": 3,
            "type": "string"
          },
          "password": {
            "maxLength": 1000,
            "minLength": 8,
            "type": "string",
            "format": "password"
          }
        },
        "additionalProperties": false
      },
      "BasicUserApiModel": {
        "required": [
          "createdAt",
          "roles",
          "userName"
        ],
        "type": "object",
        "properties": {
          "userName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ConfirmedUserCredentialsUpdateBindingModel": {
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "format": "password"
          },
          "newPassword": {
            "maxLength": 1024,
            "minLength": 8,
            "type": "string",
            "format": "password"
          }
        },
        "additionalProperties": false
      }
    }
  }
}